v5.0.0
- **File formats:**
  - (Add) Anycubic PWSZ Zip file format (#892)
  - (Add) Support for Anycubic Photon Mono 4 (pm4n) and corresponding PrusaSlicer profile
  - (Add) Support for Anycubic Photon Mono 4 Ultra (pm4u) and corresponding PrusaSlicer profile
  - (Add) Support for Anycubic Photon Mono M7 (pm7) and corresponding PrusaSlicer profile
  - (Add) Support for Anycubic Photon Mono M7 Max (pm7m) and corresponding PrusaSlicer profile
  - (Add) Support for Anycubic Photon Mono M7 Pro (pwsz) and corresponding PrusaSlicer profile
- (Add) PrusaSlicer printer: Phrozen Sonic Mighty Revo (#950)
- (Add) Litophane tool: Option to enable or disable the separation of grayscale pixels (#954)
- (Add) Exposure time finder: Re-arrange exposure text layout and allow to change it font (#955)
- (Add) Setting: Available RAM lower limit - Sets a lower limiter for the available memory RAM the program is allowed to run operations.
                   When meet the threshold, a stopping action will be queued to relief pressure and maintain the system stability.
                   In cases where is unable to pause or cancel the operation, the program will be forced to close and trigger an exception to ensure the system stability.
                   Note: This limiter will check the RAM every 2 seconds while operations are running, if you have set a very low limit there is a chance to consume more RAM in the time and cause system instability.
                   System can also reserve some RAM to prevent depletion and start to use SWAP memory, doing such with a low limit and is possible that the limiter never trigger.
                   0: Ignore the RAM limiter.
                   Default: 1 GB
- (Improvement) Change the gif animation library to a more efficient and compatible one
- (Fix) Contour traverse function was duplicating the contours, causing wrong calculations
- (Fix) Blur: Prevent stack blur from use even values
- (Upgrade) AvaloniaUI from 11.1.3 to 11.2.2
- (Upgrade) .NET from 6.0.33 to 9.0.0
  - .NET 6.0 is end of life and will no longer supported
  - This represents three major upgrades and will increase the system os version requirements, see more here: https://github.com/dotnet/core/blob/main/release-notes/9.0/supported-os.md
  - With this upgrade the software will be able to take advantage of the new features and improvements of the .NET 9.0, including a significant performance boost

+++ UVtools.Core/Operations/OperationLayerExportGif.cs
@@ -6,18 +6,22 @@
  *  of this license document, but changing it is not allowed.
  */
 
-using AnimatedGif;
 using Emgu.CV;
 using Emgu.CV.CvEnum;
 using Emgu.CV.Structure;
+using SixLabors.ImageSharp.PixelFormats;
+using SixLabors.ImageSharp;
 using System;
 using System.ComponentModel;
-using System.Drawing;
 using System.IO;
 using System.Text;
 using System.Threading.Tasks;
 using UVtools.Core.Extensions;
 using UVtools.Core.FileFormats;
+using Point = System.Drawing.Point;
+using Size = System.Drawing.Size;
+using SixLabors.ImageSharp.Formats.Gif;
+using System.Linq;
 
 namespace UVtools.Core.Operations;
 
@@ -212,9 +216,11 @@ public sealed class OperationLayerExportGif : Operation
 
     protected override bool ExecuteInternally(OperationProgress progress)
     {
-        using var gif = AnimatedGif.AnimatedGif.Create(_filePath, FPSToMilliseconds, _repeats);
-        var layerBuffer = new byte[TotalLayers][];
-        progress.Reset("Optimized layers", TotalLayers);
+
+        //using var gif = AnimatedGif.AnimatedGif.Create(_filePath, FPSToMilliseconds, _repeats);
+        // Set animation loop repeat count to 5.
+        
+        progress.Reset("Packed layers", TotalLayers);
 
         var fontFace = FontFace.HersheyDuplex;
         float fontScale = 1.5f;
@@ -226,7 +232,94 @@ public sealed class OperationLayerExportGif : Operation
             ROI = SlicerFile.BoundingRectangle;
         }
 
-        Parallel.For(0, TotalLayers, CoreSettings.GetParallelOptions(progress), i =>
+        var roiSize = GetRoiSizeOrDefault();
+        var imgSize = new Size((int)(roiSize.Width * ScaleFactor), (int)(roiSize.Height * ScaleFactor));
+        var finalSize = Size.Empty;
+
+        Image<L8>? gif = null;
+
+        var delay = FPSToMilliseconds / 10;
+        var layerBuffer = new byte[TotalLayers][];
+        var batches = Enumerable.Range(0, (int)TotalLayers).Chunk(FileFormat.DefaultParallelBatchCount);
+        foreach (var batch in batches)
+        {
+            Parallel.ForEach(batch, CoreSettings.GetParallelOptions(progress), i =>
+            {
+                progress.PauseIfRequested();
+                uint layerIndex = (uint)(LayerIndexStart + i * (_skip + 1));
+                var layer = SlicerFile[layerIndex];
+                using var mat = layer.LayerMat;
+                using var matRoi = GetRoiOrDefault(mat);
+
+                if (_scale != 100)
+                {
+                    CvInvoke.Resize(matRoi, matRoi, imgSize);
+                }
+
+                if (_flipDirection != FlipDirection.None)
+                {
+                    CvInvoke.Flip(matRoi, matRoi, (FlipType)_flipDirection);
+                }
+
+                if (_rotateDirection != RotateDirection.None)
+                {
+                    CvInvoke.Rotate(matRoi, matRoi, (RotateFlags)_rotateDirection);
+                }
+
+                if (_renderLayerCount)
+                {
+                    int baseLine = 0;
+                    var text = string.Format($"{{0:D{SlicerFile.LayerDigits}}}/{{1}}",
+                        layerIndex,
+                        SlicerFile.LayerCount - 1);
+                    var fontSize = CvInvoke.GetTextSize(text, fontFace, fontScale, fontThickness, ref baseLine);
+
+                    Point point = new(
+                        matRoi.Width / 2 - fontSize.Width / 2,
+                        70);
+                    CvInvoke.PutText(matRoi, text, point, fontFace, fontScale, textColor, fontThickness, LineType.AntiAlias);
+                }
+
+                //ApplyMask(matOriginal, matRoi);
+
+                if (finalSize == Size.Empty)
+                {
+                    finalSize = matRoi.Size;
+                }
+
+                layerBuffer[i] = matRoi.GetBytes();
+                progress.LockAndIncrement();
+            });
+
+            if (gif is null)
+            {
+                gif = new Image<L8>(imgSize.Width, imgSize.Height);
+                var gifMetaData = gif.Metadata.GetGifMetadata();
+                gifMetaData.RepeatCount = _repeats;
+
+                var metadata = gif.Frames.RootFrame.Metadata.GetGifMetadata();
+                metadata.FrameDelay = delay;
+            }
+
+            foreach (var i in batch)
+            {
+                // Create a color image, which will be added to the gif.
+                using var image = Image.LoadPixelData<L8>(layerBuffer[i], finalSize.Width, finalSize.Height);
+                layerBuffer[i] = null!;
+
+                // Set the delay until the next image is displayed.
+                var metadata = image.Frames.RootFrame.Metadata.GetGifMetadata();
+                metadata.FrameDelay = delay;
+                metadata.DisposalMethod = GifDisposalMethod.RestoreToBackground;
+
+                // Add the color image to the gif.
+                gif.Frames.AddFrame(image.Frames.RootFrame);
+
+                progress.PauseOrCancelIfRequested();
+            }
+        }
+
+        /*Parallel.For(0, TotalLayers, CoreSettings.GetParallelOptions(progress), i =>
         {
             progress.PauseIfRequested();
             uint layerIndex = (uint) (LayerIndexStart + i * (_skip + 1));
@@ -237,7 +330,7 @@ public sealed class OperationLayerExportGif : Operation
 
             if (_scale != 100)
             {
-                CvInvoke.Resize(matRoi, matRoi, new Size((int) (matRoi.Width * ScaleFactor), (int)(matRoi.Height * ScaleFactor)));
+                CvInvoke.Resize(matRoi, matRoi, imgSize);
             }
 
             if (_flipDirection != FlipDirection.None)
@@ -266,31 +359,59 @@ public sealed class OperationLayerExportGif : Operation
 
             //ApplyMask(matOriginal, matRoi);
 
+            if (finalSize == Size.Empty)
+            {
+                finalSize = matRoi.Size;
+            }
+
             layerBuffer[i] = matRoi.GetPngByes();
 
             progress.LockAndIncrement();
         });
 
+
+
         progress.ResetNameAndProcessed("Packed layers");
         foreach (var buffer in layerBuffer)
         {
             progress.PauseOrCancelIfRequested();
-            using var stream = new MemoryStream(buffer);
-            using var img = Image.FromStream(stream);
-            gif.AddFrame(img, -1, GifQuality.Bit8);
+            //using var stream = new MemoryStream(buffer);
+            //using var img = Image.FromStream(stream);
+            //gif.AddFrame(img, -1, GifQuality.Bit8);
+
+            // Create a color image, which will be added to the gif.
+            using var image = Image.Load<L8>(buffer);
+
+            // Set the delay until the next image is displayed.
+            metadata = image.Frames.RootFrame.Metadata.GetGifMetadata();
+            metadata.FrameDelay = delay;
+            metadata.DisposalMethod = GifDisposalMethod.RestoreToBackground;
+
+            // Add the color image to the gif.
+            gif.Frames.AddFrame(image.Frames.RootFrame);
+
             progress++;
         }
+        */
 
-        if (progress.Token.IsCancellationRequested)
+        progress.Reset("Saving GIF to file");
+
+        if (!progress.Token.IsCancellationRequested && gif is not null)
         {
             try
             {
+                gif.Frames.RemoveFrame(0);
+                gif.SaveAsGif(_filePath);
+            }
+            catch (Exception)
+            {
                 File.Delete(_filePath);
             }
-            catch
+            finally
             {
-                // ignored
+                gif.Dispose();
             }
+            
         }
 
         return !progress.Token.IsCancellationRequested;
